Changed the layout
settings now go in the settings overflow menu, as well as log out.
Main activity now contains a single fragment, that can be replaced as
needed. It's default value is the new Container fragment.
Split the usage panel to Usage and Fuel log.
Added interface for communication between the many fragments and Main
Activity.